ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
groq
httpx
python-dotenv
pydantic